requirements.txt
discord.py
openai>=1.2.0